refactor: split header parsing into its own file
